refactor(audit): thin command adapters, move logic to core (#5902)
