+++ src/commands/agent_task/tests/support.rs
@@ -18,7 +18,6 @@ pub(in crate::commands::agent_task) use super::super::args::{
 pub(in crate::commands::agent_task) use super::super::controller::{
     apply_controller_event, controller_from_spec, controller_materialize,
     controller_run_action_with_executor, controller_run_next_with_executor,
-    dispatch_args_from_controller_request,
 };
 pub(in crate::commands::agent_task) use super::super::run::{
     retry, run_loaded_plan, run_loop_with_executor, run_next_with_executor,
