+++ src/commands/daemon.rs
@@ -1,9 +1,6 @@
 use clap::{Args, Parser, Subcommand};
 use serde::Serialize;
 use std::path::PathBuf;
-use std::process::{Command, Stdio};
-use std::thread;
-use std::time::{Duration, Instant};
 
 use homeboy::core::daemon::{
     self, BrokerConfig, BrokerConfigOptions, DaemonStartResult, DaemonStatus, DaemonStopResult,
@@ -99,7 +96,9 @@ pub struct DaemonArtifactGetOutput {
 
 pub fn run(args: DaemonArgs, _global: &crate::commands::GlobalArgs) -> CmdResult<DaemonOutput> {
     match args.command {
-        DaemonCommand::Start { addr } => start(&addr),
+        DaemonCommand::Start { addr } => {
+            Ok((DaemonOutput::Start(daemon::start_background(&addr)?), 0))
+        }
         DaemonCommand::Serve { addr } => serve(&addr),
         DaemonCommand::Stop => Ok((DaemonOutput::Stop(daemon::stop()?), 0)),
         DaemonCommand::Status => Ok((DaemonOutput::Status(daemon::read_status()?), 0)),
@@ -126,118 +125,27 @@ pub fn run(args: DaemonArgs, _global: &crate::commands::GlobalArgs) -> CmdResult
 }
 
 fn artifact_get(args: DaemonArtifactGetArgs) -> CmdResult<DaemonOutput> {
-    let daemon_url = resolve_daemon_url(args.daemon_url)?;
-    let content_url = artifact_content_url(&daemon_url, &args.run_id, &args.artifact_id)?;
-    let output_path = args
-        .output
-        .unwrap_or_else(|| default_artifact_output_path(&args.artifact_id));
-    let response = reqwest::blocking::get(&content_url).map_err(reqwest_error)?;
-    let status = response.status();
-    let headers = response.headers().clone();
-    if !status.is_success() {
-        let body = response.text().unwrap_or_default();
-        return Err(homeboy::core::Error::validation_invalid_argument(
-            "artifact_id",
-            format!(
-                "daemon artifact fetch failed with HTTP {}: {}",
-                status.as_u16(),
-                body
-            ),
-            Some(args.artifact_id),
-            None,
-        ));
-    }
-
-    let bytes = response.bytes().map_err(reqwest_error)?;
-    if let Some(parent) = output_path
-        .parent()
-        .filter(|parent| !parent.as_os_str().is_empty())
-    {
-        std::fs::create_dir_all(parent).map_err(|e| {
-            homeboy::core::Error::internal_io(
-                e.to_string(),
-                Some(format!("create {}", parent.display())),
-            )
-        })?;
-    }
-    std::fs::write(&output_path, &bytes).map_err(|e| {
-        homeboy::core::Error::internal_io(
-            e.to_string(),
-            Some(format!("write {}", output_path.display())),
-        )
-    })?;
+    let outcome = daemon::fetch_artifact_to_path(
+        &args.run_id,
+        &args.artifact_id,
+        args.daemon_url,
+        args.output,
+    )?;
 
     Ok((
         DaemonOutput::ArtifactGet(DaemonArtifactGetOutput {
             command: "daemon.artifact.get",
             run_id: args.run_id,
             artifact_id: args.artifact_id,
-            daemon_url,
-            content_url,
-            output_path: output_path.display().to_string(),
-            content_type: header_value(&headers, reqwest::header::CONTENT_TYPE.as_str()),
-            size_bytes: Some(bytes.len() as u64),
-            sha256: header_value(&headers, "x-homeboy-artifact-sha256"),
+            daemon_url: outcome.daemon_url,
+            content_url: outcome.content_url,
+            output_path: outcome.output_path.display().to_string(),
+            content_type: outcome.content_type,
+            size_bytes: Some(outcome.size_bytes),
+            sha256: outcome.sha256,
         }),
         0,
     ))
-}
-
-fn resolve_daemon_url(daemon_url: Option<String>) -> homeboy::core::Result<String> {
-    if let Some(url) = daemon_url.filter(|url| !url.trim().is_empty()) {
-        return Ok(url);
-    }
-    let status = daemon::read_status()?;
-    let Some(state) = status.state.filter(|_| status.running) else {
-        return Err(homeboy::core::Error::validation_invalid_argument(
-            "daemon-url",
-            "daemon is not running; pass --daemon-url or start it with `homeboy daemon start`",
-            None,
-            None,
-        ));
-    };
-    Ok(format!("http://{}", state.address))
-}
-
-fn artifact_content_url(
-    daemon_url: &str,
-    run_id: &str,
-    artifact_id: &str,
-) -> homeboy::core::Result<String> {
-    let mut base = reqwest::Url::parse(daemon_url).map_err(|e| {
-        homeboy::core::Error::validation_invalid_argument(
-            "daemon-url",
-            e.to_string(),
-            Some(daemon_url.to_string()),
-            None,
-        )
-    })?;
-    base.set_path(&format!(
-        "/runs/{}/artifacts/{}/content",
-        homeboy::core::execution_contract::encode_uri_component(run_id),
-        homeboy::core::execution_contract::encode_uri_component(artifact_id)
-    ));
-    base.set_query(None);
-    Ok(base.to_string())
-}
-
-fn default_artifact_output_path(artifact_id: &str) -> PathBuf {
-    artifact_id
-        .rsplit('/')
-        .find(|segment| !segment.is_empty())
-        .map(PathBuf::from)
-        .unwrap_or_else(|| PathBuf::from("artifact.bin"))
-}
-
-fn header_value(headers: &reqwest::header::HeaderMap, name: &str) -> Option<String> {
-    headers
-        .get(name)
-        .and_then(|value| value.to_str().ok())
-        .map(str::to_string)
-}
-
-fn reqwest_error(error: reqwest::Error) -> homeboy::core::Error {
-    homeboy::core::Error::internal_io(error.to_string(), Some("fetch daemon artifact".to_string()))
 }
 
 fn serve(addr: &str) -> CmdResult<DaemonOutput> {
@@ -251,53 +159,6 @@ fn serve(addr: &str) -> CmdResult<DaemonOutput> {
         }),
         0,
     ))
-}
-
-fn start(addr: &str) -> CmdResult<DaemonOutput> {
-    daemon::parse_bind_addr(addr)?;
-
-    let exe = std::env::current_exe().map_err(|e| {
-        homeboy::core::Error::internal_io(
-            e.to_string(),
-            Some("resolve current executable".to_string()),
-        )
-    })?;
-    let child = Command::new(exe)
-        .args(["daemon", "serve", "--addr", addr])
-        .stdin(Stdio::null())
-        .stdout(Stdio::null())
-        .stderr(Stdio::null())
-        .spawn()
-        .map_err(|e| {
-            homeboy::core::Error::internal_io(e.to_string(), Some("spawn daemon".to_string()))
-        })?;
-    let pid = child.id();
-
-    let deadline = Instant::now() + Duration::from_secs(5);
-    loop {
-        let status = daemon::read_status()?;
-        if let Some(state) = status.state {
-            if state.pid == pid {
-                return Ok((
-                    DaemonOutput::Start(DaemonStartResult {
-                        pid,
-                        address: state.address,
-                        state_path: state.state_path,
-                    }),
-                    0,
-                ));
-            }
-        }
-
-        if Instant::now() >= deadline {
-            return Err(homeboy::core::Error::internal_unexpected(format!(
-                "daemon process {} did not publish state before timeout",
-                pid
-            )));
-        }
-
-        thread::sleep(Duration::from_millis(50));
-    }
 }
 
 #[derive(Debug, Clone, Copy)]
@@ -332,21 +193,6 @@ mod tests {
     use homeboy::test_support::with_isolated_home;
 
     use super::*;
-
-    #[test]
-    fn artifact_content_url_builds_encoded_daemon_byte_alias() {
-        let url = artifact_content_url(
-            "http://127.0.0.1:7421/base?ignored=true",
-            "run 1",
-            "report/summary.json",
-        )
-        .expect("url");
-
-        assert_eq!(
-            url,
-            "http://127.0.0.1:7421/runs/run%201/artifacts/report%2Fsummary.json/content"
-        );
-    }
 
     #[test]
     fn artifact_get_downloads_daemon_byte_alias() {
