+++ src/core/mod.rs
@@ -50,6 +50,7 @@ pub(crate) mod artifact_metadata;
 pub mod artifact_origin;
 pub mod artifact_ref;
 pub mod browser_evidence;
+pub mod browser_visual_compare;
 pub mod build_identity;
 pub mod change_artifact;
 pub mod ci_failure_log_triage;
